GCodeProcessor final update
- g_code was changed to an array[dictionary] format
- this enables a better handling of the g_code inside and outside of the processor
